requirements.txt
feedparser
openai
pywhatkit
pyautogui
google-generativeai